requirements.txt
streamlit
openai
pyyaml
python-docx
langchain
python-dotenv
urllib3==1.26.6
Werkzeug==2.0.1
